Moved EncodingTestType, TestEncoders and TestDecoders to quill-core.
