+++ quill-async-mysql/src/test/scala/io/getquill/context/async/mysql/TestContext.scala
@@ -1,6 +1,5 @@
 package io.getquill.context.async.mysql
 
-import io.getquill.{ Literal, MysqlAsyncContext, TestEntities }
-import io.getquill.context.sql.{ TestDecoders, TestEncoders }
+import io.getquill._
 
-class TestContext extends MysqlAsyncContext[Literal]("testMysqlDB") with TestEntities with TestEncoders with TestDecoders
+class TestContext extends MysqlAsyncContext[Literal]("testMysqlDB") with TestExtras
