+++ quill-cassandra/src/test/scala/io/getquill/context/cassandra/package.scala
@@ -1,22 +1,19 @@
 package io.getquill.context
 
-import scala.concurrent.Await
-import scala.concurrent.Future
-import scala.concurrent.duration.Duration
-
 import io.getquill._
-import io.getquill.Literal
-import io.getquill.CassandraStreamContext
+
+import scala.concurrent.duration.Duration
+import scala.concurrent.{ Await, Future }
 
 package object cassandra {
 
-  lazy val mirrorContext = new CassandraMirrorContext with TestEntities
+  lazy val mirrorContext = new CassandraMirrorContext with TestExtras
 
-  lazy val testSyncDB = new CassandraSyncContext[Literal]("testSyncDB") with TestEntities
+  lazy val testSyncDB = new CassandraSyncContext[Literal]("testSyncDB") with TestExtras
 
-  lazy val testAsyncDB = new CassandraAsyncContext[Literal]("testAsyncDB") with TestEntities
+  lazy val testAsyncDB = new CassandraAsyncContext[Literal]("testAsyncDB") with TestExtras
 
-  lazy val testStreamDB = new CassandraStreamContext[Literal]("testStreamDB") with TestEntities
+  lazy val testStreamDB = new CassandraStreamContext[Literal]("testStreamDB") with TestExtras
 
   def await[T](f: Future[T]): T = Await.result(f, Duration.Inf)
 }
