+++ quill-core/src/test/scala/io/getquill/EncodingTestType.scala
@@ -1,0 +1,3 @@
+package io.getquill
+
+case class EncodingTestType(value: String)
