+++ quill-core/src/test/scala/io/getquill/TestDecoders.scala
@@ -1,6 +1,4 @@
-package io.getquill.context.sql
-
-import io.getquill.MappedEncoding
+package io.getquill
 
 trait TestDecoders {
   implicit val encodingTestTypeDecoder = MappedEncoding[String, EncodingTestType](EncodingTestType)
