+++ quill-core/src/test/scala/io/getquill/TestEncoders.scala
@@ -1,6 +1,4 @@
-package io.getquill.context.sql
-
-import io.getquill.MappedEncoding
+package io.getquill
 
 trait TestEncoders {
   implicit val encodingTestTypeEncoder = MappedEncoding[EncodingTestType, String](_.value)
