+++ quill-core/src/test/scala/io/getquill/TestExtras.scala
@@ -1,0 +1,6 @@
+package io.getquill
+
+trait TestExtras extends TestEntities with TestEncoders with TestDecoders {
+  this: context.Context[_, _] =>
+}
+
