+++ quill-finagle-mysql/src/test/scala/io/getquill/context/finagle/mysql/TestContext.scala
@@ -1,8 +1,5 @@
 package io.getquill.context.finagle.mysql
 
-import io.getquill.Literal
-import io.getquill.TestEntities
-import io.getquill.FinagleMysqlContext
-import io.getquill.context.sql.{ TestDecoders, TestEncoders }
+import io.getquill.{FinagleMysqlContext, Literal, TestExtras}
 
-object testContext extends FinagleMysqlContext[Literal]("testDB") with TestEntities with TestEncoders with TestDecoders
+object testContext extends FinagleMysqlContext[Literal]("testDB") with TestExtras
