+++ quill-jdbc/src/test/scala/io/getquill/context/jdbc/mysql/TestContext.scala
@@ -1,7 +1,5 @@
 package io.getquill.context.jdbc.mysql
 
-import io.getquill.context.sql.TestEncoders
-import io.getquill.context.sql.TestDecoders
-import io.getquill.{ JdbcContext, Literal, MySQLDialect, TestEntities }
+import io.getquill.{JdbcContext, Literal, MySQLDialect, TestExtras}
 
-object testContext extends JdbcContext[MySQLDialect, Literal]("testMysqlDB") with TestEntities with TestEncoders with TestDecoders
+object testContext extends JdbcContext[MySQLDialect, Literal]("testMysqlDB") with TestExtras
