+++ quill-jdbc/src/test/scala/io/getquill/context/jdbc/postgres/TestContext.scala
@@ -1,7 +1,6 @@
 package io.getquill.context.jdbc.postgres
 
-import io.getquill.{ JdbcContext, Literal, PostgresDialect, TestEntities }
-import io.getquill.context.sql.TestEncoders
-import io.getquill.context.sql.TestDecoders
+import io.getquill.{JdbcContext, Literal, PostgresDialect, TestExtras}
 
-object testContext extends JdbcContext[PostgresDialect, Literal]("testPostgresDB") with TestEntities with TestEncoders with TestDecoders
+
+object testContext extends JdbcContext[PostgresDialect, Literal]("testPostgresDB") with TestExtras
