+++ quill-jdbc/src/test/scala/io/getquill/context/jdbc/sqlite/TestContext.scala
@@ -1,7 +1,5 @@
 package io.getquill.context.jdbc.sqlite
 
-import io.getquill.{ JdbcContext, Literal, SqliteDialect, TestEntities }
-import io.getquill.context.sql.TestEncoders
-import io.getquill.context.sql.TestDecoders
+import io.getquill.{JdbcContext, Literal, SqliteDialect, TestExtras}
 
-object testContext extends JdbcContext[SqliteDialect, Literal]("testSqliteDB") with TestEntities with TestEncoders with TestDecoders
+object testContext extends JdbcContext[SqliteDialect, Literal]("testSqliteDB") with TestExtras
